# doc/openapi.yml
openapi: 3.0.1
info:
  title: JobBank1111.Job.WebAPI
  version: '1.0'
paths:
...
  /api/v2/members:
    post:
      tags:
        - Member
      operationId: InsertMember1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertMemberRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/InsertMemberRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/InsertMemberRequest'
      responses:
        '200':
          description: OK
components:
  schemas:
...
    InsertMemberRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        age:
          type: integer
          format: int32
      additionalProperties: false
...